requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=1.3.0
keras>=2.1.6
opencv-python
h5py
imgaug
IPython[all]